move Validator base class to validators package
