+++ src/main/java/gedcom/validators/LessThan150YearsOld.java
@@ -1,6 +1,5 @@
 package gedcom.validators;
 
-import gedcom.Validator;
 import gedcom.models.GEDFile;
 import gedcom.models.Individual;
